+++ trunk/RTSPProxy-Core/src/main/java/rtspproxy/lib/number/UnsignedInt.java
@@ -82,12 +82,12 @@ public final class UnsignedInt extends UnsignedNumber {
 
 	@Override
 	public double doubleValue() {
-		return (double) value;
+		return value;
 	}
 
 	@Override
 	public float floatValue() {
-		return (float) value;
+		return value;
 	}
 
 	@Override
@@ -129,7 +129,7 @@ public final class UnsignedInt extends UnsignedNumber {
 
 	@Override
 	public String toString() {
-		return Long.toString((long) value & 0xFFFFFFFFL);
+		return Long.toString( value & 0xFFFFFFFFL);
 	}
 	
 	@Override
